Finalizada a rotina que pesquisa pelo CEP e armazena na base de dados local.
