+++ resources/views/admin/person/create.blade.php
@@ -11,7 +11,7 @@
         <div class="card-header">
             <h3 class="card-title"></h3>
         </div>
-        <form action="{{ route(strtolower($model) . '.store') }}" method="post" onsubmit="return false;">
+        <form action="{{ route(strtolower($model) . '.store') }}" method="post">
             @include('admin.person.partials.form', ['model' => $model])
         </form>
     </div>
@@ -22,33 +22,4 @@
 
 @section('js')
     @vite(['resources/js/person.js'])
-    <script>
-        $('#btnZipCodeSearch').on('click', function () {
-            $.ajax({
-                url: '/zipcode/search',
-                type: "get",
-                dataType: 'json',
-                data: {
-                    'zip_code': $('#zip_code').inputmask('unmaskedvalue')
-                },
-            }).done(function(data) {
-                if (!('erro' in data)) {
-                    $('#street').val(data.street);
-                    $('#district').val(data.district);
-                    exibirUF(data.state);
-                    exibirCidade(null);
-                }
-            });
-        });
-
-        function exibirUF(state)
-        {
-            $('#state').append($('<option>', {value: state.id, text: state.name}));
-        }
-
-        function exibirCidade(city)
-        {
-            $('#city').prop("disabled", false);
-        }
-    </script>
 @stop
