+++ resources/views/admin/person/edit.blade.php
@@ -7,7 +7,7 @@
 @stop
 
 @section('content')
-    <form action="{{ route(strtolower($model) . '.update', $person->id) }}" method="post" onsubmit="return false;">
+    <form action="{{ route(strtolower($model) . '.update', $person->id) }}" method="post">
         @method('PUT')
         @include('admin.person.partials.form', ['person' => $person, 'model' => $model])
     </form>
